fix(remote): allows the `ApproveButton` component to get its port number from the browser
currently this component is trying to read from `process.env.BACKSTOP_REMOTE_HTTP_PORT`, which the browser has no access to. not sure if th
e intent is for webpack to interpolate this value and output a string, but it currently isn't. let the browser do the work for you :)